pyproject.toml
[project]
name = "beyond-chatgpt"
version = "0.1.0"
description = "gpt-like-a-dev"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "fastapi>=0.115.12",
    "jupyter>=1.1.1",
    "boto3>=1.35.0",
    "pydantic>=2.11.4",
    "uvicorn>=0.34.2",
    "python-multipart>=0.0.18",
    "python-dotenv>=1.0.0",
]
